1ère version de useRules qui reprend RulesProvider, sur Model.tsx
