+++ source/RulesProvider.js
@@ -10,97 +10,18 @@ import {
 import useBranchData from 'Components/useBranchData'
 import Engine from 'publicodes'
 import { useEffect, useMemo } from 'react'
-import { Trans, useTranslation } from 'react-i18next'
+import { Trans } from 'react-i18next'
 import { useDispatch, useSelector } from 'react-redux'
 
 import { constantFolding, getRawNodes } from 'publiopti'
-import { addTranslationToBaseRules } from '../nosgestesclimat/scripts/i18n/addTranslationToBaseRules'
-import { getCurrentLangAbrv } from './locales/translation'
-
-/* This component gets the publicode rules from the good URL,
- * then gives them
- * to the engine to parse, and hence makes it available to the whole component tree
- * through the state (state.rules) as unparsed, or through the useEngine hook as parsed, but only for component that are enclosed in WithEngine
- * to trigger the parsing only for components that need this heavy operation.
- *
- * This component triggers loading rules as soon as possible, BUT the components that use
- * the engine should wait for it to be available. Hence the use of the WithRules
- * component that returns null if rules are not ready in the state.
- *
- * This logic is a handmade and basic implementation of react 18's Suspense for data loading
- * principles. Switching to this experimental feature could be great if we had concurrent
- * loading problems. Here, we only have one block of data (co2.json) at a time.
- * */
+import useRules from './components/useRules'
 
 export default ({ children }) => {
-	const { i18n } = useTranslation()
-	const currLangAbrv = getCurrentLangAbrv(i18n)
-	const branchData = useBranchData()
-	const rules = useSelector((state) => state.rules)
-
-	const dispatch = useDispatch()
-
-	const setRules = (rules) => dispatch({ type: 'SET_RULES', rules })
-
-	useEffect(() => {
-		if (!branchData.loaded) return
-		//This NODE_ENV condition has to be repeated here, for webpack when compiling. It can't interpret shouldUseLocalFiles even if it contains the same variable
-		if (NODE_ENV === 'development' && branchData.shouldUseLocalFiles) {
-			// TODO: find a way to use compressed models in dev mode
-			console.log(
-				'===== DEV MODE : the model is on your hard drive on ./nosgestesclimat ======='
-			)
-			// Rules are stored in nested yaml files
-			const req = require.context('../nosgestesclimat/data/', true, /\.(yaml)$/)
-
-			const baseRules = req.keys().reduce((acc, key) => {
-				if (key.match(/translated-rules-.*yaml/)) {
-					// ignoring translating files.
-					return acc
-				}
-				const jsonRuleSet = req(key).default || {}
-				return { ...acc, ...jsonRuleSet }
-			}, {})
-
-			var rules = baseRules
-
-			const currentLang = i18n.language === 'en' ? 'en-us' : i18n.language
-			if (currentLang !== 'fr') {
-				const translatedRulesAttrs =
-					require(`../nosgestesclimat/data/translated-rules-${currentLang}.yaml`).default
-				rules = addTranslationToBaseRules(baseRules, translatedRulesAttrs)
-				if (!rules) {
-					console.error(
-						'Error occured while recompiling translated rules for:',
-						currentLang
-					)
-				}
-			}
-
-			setRules(rules, branchData.deployURL)
-		} else {
-			const url =
-				branchData.deployURL +
-				// TODO: find a better way to manage 'en'
-				`/co2-${i18n.language === 'en' ? 'en-us' : currLangAbrv}-opti.json`
-			console.log('fetching:', url)
-			fetch(url, { mode: 'cors' })
-				.then((response) => response.json())
-				.then((json) => {
-					setRules(json, branchData.deployURL)
-				})
-		}
-	}, [
-		branchData.deployURL,
-		branchData.loaded,
-		branchData.shouldUseLocalFiles,
-		i18n.language,
-	])
-
-	return <EngineWrapper rules={rules}>{children}</EngineWrapper>
+	return <EngineWrapper>{children}</EngineWrapper>
 }
 
-const EngineWrapper = ({ rules, children }) => {
+const EngineWrapper = ({ children }) => {
+	const rules = useRules()
 	const engineState = useSelector((state) => state.engineState)
 	const dispatch = useDispatch()
 	const branchData = useBranchData()
